requirements.txt
flask_marshmallow==0.14.0
Flask==2.0.1
alembic==1.6.5
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.17
python-dotenv==0.19.1
